requirements.txt
pyyaml>=6.0
types-PyYAML>=6.0
typing-extensions>=4.5
